Adicionamento de metodos nas classes

+++ ProjetoHospitalClassAtributos/Consulta.cs
@@ -7,12 +7,12 @@ namespace ProjetoHospitalClassAtributos
 {
     public class Consulta
     {
-        private byte id;
+        private int id;
         private string data;
         private string hora;
         private string descricao;
 
-        public byte Id{get;set;}
+        public int Id{get;set;}
         public string Data{get;set;}
         public string Hora{get;set;}
         public string Descricao{get;set;}
